+++ Cell-Tribe/Entity/EntityLiving.h
@@ -8,7 +8,7 @@ public:
 	virtual int update() override final;
 	virtual int print() const override;
 	virtual bool canBeAttacked() const override final;
-	virtual int beAttacked(EntityLiving*, const LL&) override;
+	virtual int beAttacked(EntityLiving*, const double&) override;
 	virtual void heal(const LL&);
 protected:
 	double atk, healthMax;
